lambdas/pricing_sheet/requirements.txt
boto3==1.28.39
openpyxl==3.1.2
psycopg2-binary==2.9.7
